Move generic classes into separate package
