+++ src/main/java/org/entur/gbfs/http/HttpUtils.java
@@ -1,4 +1,4 @@
-package org.entur.gbfs;
+package org.entur.gbfs.http;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
